docs: move gatsby config back to js
`ts-node` was slowing down startup time.
