+++ .gitignore
@@ -137,6 +137,8 @@ secret.json
 .out
 
 docs/public/**
+docs/.cache/
+docs/.linaria-cache/
 *.tsbuildinfo
 e2e/__failed-diffs__
 cc-reporter
