+++ @remirror/playground/package.json
@@ -35,8 +35,10 @@
     "remirror": "^0.13.1"
   },
   "devDependencies": {
+    "@types/prettier": "^2.0.1",
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
+    "prettier": "^2.0.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "ts-node": "^8.10.2"
