--- docs/gatsby-node.ts
@@ -1,6 +1,0 @@
-export {
-  createPages,
-  onCreateBabelConfig,
-  onCreateNode,
-  onCreateWebpackConfig,
-} from './src/gatsby/node';
