+++ docs/playground.mdx
@@ -3,9 +3,6 @@ title: Playground
 fullWidth: true
 ---
 
-import * as foo from '@remirror/playground';
-import '@remirror/styles/all.css';
+import { Playground } from '@remirror/playground';
 
-<foo.Playground />
-
-<p>{((React.Component._hello = 123), 'done')}</p>
+<Playground />
