+++ docs/showcase/social.mdx
@@ -2,8 +2,6 @@
 title: Social Editor Showcase
 ---
 
-import '@remirror/styles/all.css';
-
 import { ExampleSocialEditor } from '@remirror/react-showcase';
 
 # Social Editor
