+++ docs/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../support/tsconfig.base.json",
   "compilerOptions": {
-    "module": "commonjs",
-    "target": "es5",
+    "allowJs": true,
+    "checkJs": true,
     "lib": ["dom", "esnext"],
     "jsx": "react"
   },
