+++ tsconfig.lint.json
@@ -4,5 +4,5 @@
     "allowJs": true
   },
   "include": ["**/*.ts", "**/*.tsx", "**/*.js", "*.js", ".eslintrc.js", "**/.babelrc.js"],
-  "exclude": ["**/dist/", "**/public/", "**/.cache/", "docs/**/*.js"]
+  "exclude": ["**/dist/", "**/public/", "**/.cache/", "**/.linaria-cache/"]
 }
